parity: five battery finds cross — the email-as-uuid insert, the wire shapes, and the decay interval
- research: the step-time research_runs insert bound requested_by (a human
  starter's email) as $3::uuid — every human-started run died at its first
  step. The route-time insert never had the cast; TS has none either.
- home: OrgGlance serialized cost_today snake_cased (TS: costToday), and
  queue items dropped the `queue` bucket TS's rows carry undeclared.
- uploads: the Multipart extractor rejected bad bodies before the permission
  gate could run, answering axum's own 400 text; the rejection is now
  captured so the gate orders first and the body reads {"error":"no file"}.
- comms_decay: make_interval(days => $1) took the i64 bind as bigint and
  matched no overload; $1::int is the cast PG wants (TS inlined the literal).

Battery state after: 370/370 pairs accounted — the remaining mcp/library
body drift is the live official registry rotating between draws, and the
google callbacks redirect to identical locations (the sweep followed them
into SvelteKit-only paths).

+++ api/src/comms_decay.rs
@@ -316,7 +316,7 @@ pub async fn sweep_idle_chats(deps: &DecayDeps) -> Result<DecaySweepResult, Stri
     let idle: Vec<IdleConv> = sqlx::query_as(
         "select id::text, user_id::text, agent_model, title from conversations \
          where kind = 'chat' and archived = false \
-           and updated_at < now() - make_interval(days => $1) \
+           and updated_at < now() - make_interval(days => $1::int) \
          order by updated_at asc limit $2",
     )
     .bind(ttl_days())
